+++ src/omni_epd/displayfactory.py
@@ -22,11 +22,11 @@ import configparser
 import importlib
 import os
 import logging
-from . errors import EPDNotFoundError
+from . errors import EPDNotFoundError, EPDConfigurationError
 from . conf import CONFIG_FILE, EPD_CONFIG
 from . virtualepd import VirtualEPD
 from . displays.mock_display import MockDisplay  # noqa: F401
-from . displays.waveshare_display import WaveshareDisplay, WaveshareTriColorDisplay, Waveshare102inDisplay, Waveshare3in7Display  # noqa: F401
+from . displays.waveshare_display import WaveshareDisplay, WaveshareTriColorDisplay, Waveshare102inDisplay, WaveshareGrayscaleDisplay, WaveshareMultiColorDisplay  # noqa: F401,E501
 from . displays.inky_display import InkyDisplay, InkyImpressionDisplay  # noqa: F401
 
 
@@ -97,6 +97,11 @@ def load_display_driver(displayName='', configDict={}):
         classObj = getattr(mod, foundClass[0]['class'])
 
         result = classObj(deviceType[1], config)
+
+        # check that the display mode is valid - must be done after class loaded
+        if(result.mode not in result.modes_available):
+            raise EPDConfigurationError(displayName, "mode", result.mode)
+
     else:
         # we have a problem
         raise EPDNotFoundError(displayName)