+++ src/omni_epd/displays/inky_display.py
@@ -18,7 +18,6 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 """
 
-from .. conf import IMAGE_ENHANCEMENTS
 from .. virtualepd import VirtualEPD
 
 
@@ -29,17 +28,11 @@ class InkyDisplay(VirtualEPD):
     """
 
     pkg_name = 'inky'
+    mode = "black"  # default mode is black
+    modes_available = ("black")
 
     def __init__(self, deviceName, config):
-        super(InkyDisplay, self).__init__(deviceName, config)
-
-        # if color does not exist, add it
-        if(not self._config.has_option(IMAGE_ENHANCEMENTS, "color")):
-            if(not self._config.has_section(IMAGE_ENHANCEMENTS)):
-                self._config.add_section(IMAGE_ENHANCEMENTS)
-
-            # by default use black/white images
-            self._config.set(IMAGE_ENHANCEMENTS, "color", "1")
+        super().__init__(deviceName, config)
 
         # need to figure out what type of device we have
         dType, dColor = deviceName.split('_')
@@ -53,6 +46,18 @@ class InkyDisplay(VirtualEPD):
         elif(dType == 'what'):
             deviceObj = self.load_display_driver(self.pkg_name, 'what')
             self._device = deviceObj.InkyWHAT(dColor)
+
+        # set mode to black + any other color supported
+        if(self.mode != "black"):
+            self.modes_available = ('black', dColor)
+
+        # phat and what devices expect colors in the order white, black, other
+        if(self.mode == "red" and dColor == "red"):
+            self.palette_filter.append([255, 0, 0])
+            self.max_colors = 3
+        elif(self.mode == "yellow" and dColor == "yellow"):
+            self.palette_filter.append([255, 255, 0])
+            self.max_colors = 3
 
         # set the width and height
         self.width = self._device.width
@@ -74,16 +79,19 @@ class InkyDisplay(VirtualEPD):
         return result
 
     def _display(self, image):
+        # apply any needed conversions to this image based on the mode
+        image = self._filterImage(image)
+
+        # set the image and display
+        self._device.set_border(getattr(self._device, self._get_device_option('border', '').upper(), self._device.border_colour))
         self._device.set_image(image)
         self._device.show()
 
     def clear(self):
-        from inky import WHITE
-
         for _ in range(2):
             for y in range(self.height - 1):
                 for x in range(self.width - 1):
-                    self._device.set_pixel(x, y, WHITE)
+                    self._device.set_pixel(x, y, self._device.WHITE)
 
         self._device.show()
 
@@ -96,9 +104,12 @@ class InkyImpressionDisplay(VirtualEPD):
     """
 
     pkg_name = 'inky'
+    mode = 'color'  # this uses color by default
+    max_colors = 8  # 7 + CLEAN (no color)
+    modes_available = ('bw', 'color')
 
     def __init__(self, deviceName, config):
-        super(InkyDisplay, self).__init__(deviceName, config)
+        super().__init__(deviceName, config)
 
         # load the device driver
         deviceObj = self.load_display_driver(self.pkg_name, 'inky_uc8159')
@@ -108,6 +119,9 @@ class InkyImpressionDisplay(VirtualEPD):
         self.width = self._device.width
         self.height = self._device.height
 
+        # get colors from the inky lib (won't be used normally as inky does conversion)
+        self.palette_filter = deviceObj.DESATURATED_PALETTE
+
     @staticmethod
     def get_supported_devices():
         result = []
@@ -115,20 +129,24 @@ class InkyImpressionDisplay(VirtualEPD):
         # python libs for this might not be installed - that's ok, return nothing
         if(InkyImpressionDisplay.check_module_installed('inky')):
             # if passed return list of devices
-            result = [f"{InkyDisplay.pkg_name}.impression"]
+            result = [f"{InkyImpressionDisplay.pkg_name}.impression"]
 
         return result
 
     def _display(self, image):
-        self._device.set_image(image)
+
+        # no palette adjustments when color as the Inky lib does them from the image
+        if(self.mode == 'bw'):
+            image = self._filterImage(image)
+
+        self._device.set_border(getattr(self._device, self._get_device_option('border', '').upper(), self._device.border_colour))
+        self._device.set_image(image.convert("RGB"), saturation=self._getfloat_device_option('saturation', .5))  # .5 is default from Inky lib
         self._device.show()
 
     def clear(self):
-        from inky.inky_uc8159 import CLEAN
-
         for _ in range(2):
             for y in range(self.height - 1):
                 for x in range(self.width - 1):
-                    self._device.set_pixel(x, y, CLEAN)
+                    self._device.set_pixel(x, y, self._device.CLEAN)
 
         self._device.show()