+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## Unreleased
+
+- added device specific modes to INI file
+- removed Image Enhancements from INI having to do with colors, moved to device specific configurations
+- updated device types to use multiple color modes when available
+- updated device table to show available device modes for each supported typo
+- Inky Impression is tested - thanks @missionfloyd
+- dynamically load class files instead of using import where possible
+
 ## Version 0.1.7
 
 ### Added